Crate rspec

Modules

context
The Context module holds all the functionality for the test declaration, that is: describe, before, after, it and their variants.

events
Events are send by the Runner to signal the progression in the test suite, with the results

formatter
The formatter module expose different formatters that can be attached to the Runner and display the progression of the run under various forms

runner
The Runner is where all the examples are actually executed.