Struct rspec::context::Context

pub struct Context<'a> {
    pub tests: Vec<Testable<'a>>,
    pub before_each: Vec<Box<BeforeFunction<'a>>>,
    pub after_each: Vec<Box<AfterFunction<'a>>>,
}

A Context holds a collection of tests, a collection of closures to call before running any tests, and a collection of closure to call after all the tests..

This is effectively the struct we fill when calling ctx.it()

Fields

tests

before_each

after_each

Methods

impl<'a> Context<'a>

fn describe<F>(&mut self, name: &'a str, body: F) where F: 'a + Send + Sync + FnMut(&mut Context<'a>) -> ()

Open and name a new example group, which will be keeped as a child context of the current context.

Note that the order of declaration is respected for running the tests.

Examples

use rspec::context::rdescribe;

// `rdescribe` instanciate a runner and run it transparently
rdescribe("inside this describe, we use the context", |ctx| {

    ctx.it("should run first", || Ok(()));

    ctx.describe("open describe", |ctx| {

        ctx.it("should run second", || Ok(()));

        ctx.describe("in a describe", |ctx| {

            ctx.describe("in a describe", |_ctx| {});

            ctx.it("should run third", || Ok(()));

        });
    });

    ctx.it("should run last", || Ok(()));
});

fn it<F>(&mut self, name: &'a str, body: F) where F: 'a + Send + Sync + FnMut() -> TestResult

Register and name a closure as an example

Examples

use rspec::context::rdescribe;

// `rdescribe` instanciate a runner and run it transparently
rdescribe("inside this describe, we use the context", |ctx| {

    ctx.it("test at the root", || Ok(()));

    ctx.describe("a group of examples", |ctx| {

        ctx.it("should be usable inside a describe", || Ok(()));

        ctx.describe("a nested describe", |ctx| {

            ctx.it("should be usabel inside multiple describes", || Ok(()));
            ctx.it("should be able to declare multiple tests", || Ok(()));

        });

        ctx.it("doesn't care if it's before or after a describe", || Ok(()));
    });
});

fn before<F>(&mut self, body: F) where F: 'a + Send + Sync + FnMut() -> ()

Declares a closure that will be executed before each test of the current Context.

PLEASE NOTE: due to a bug in current versions of rspec, the before closures WILL BE CALLED ONLY ONCE for all the children of the current context.

Examples

use rspec::context::rdescribe;
use std::sync::atomic::{AtomicUsize, Ordering};

let counter = &mut AtomicUsize::new(0);

// `rdescribe` instanciate a runner and run it transparently
rdescribe("inside this describe, we use the context", |ctx| {

    // This will increment the counter at each test
    ctx.before(|| { counter.fetch_add(1, Ordering::SeqCst); });

    ctx.it("should run after the before", || {
        assert_eq!(1, counter.load(Ordering::SeqCst));
        Ok(())
    });

    ctx.describe("a group of examples", |ctx| {

        ctx.it("should see 1 increment", || {
            assert_eq!(2, counter.load(Ordering::SeqCst));
            Ok(())
        });

        // XXX - note that the before has not been applied another time
        ctx.it("should NOT see another increment", || {
            assert_eq!(2, counter.load(Ordering::SeqCst));
            Ok(())
        });
    });

    ctx.it("should run after the all the befores AND the previous it", || {
        assert_eq!(3, counter.load(Ordering::SeqCst));
        Ok(())
    });
});

fn after<F>(&mut self, body: F) where F: 'a + Send + Sync + FnMut() -> ()

Declares a closure that will be executed after each test of the current Context.

PLEASE NOTE: due to a bug in current versions of rspec, the after closures WILL BE CALLED ONLY ONCE for all the children of the current context.

Examples

use rspec::context::rdescribe;
use std::sync::atomic::{AtomicUsize, Ordering};

let counter = &mut AtomicUsize::new(0);

// `rdescribe` instanciate a runner and run it transparently
rdescribe("inside this describe, we use the context", |ctx| {

    // This will increment the counter at each test
    ctx.after(|| { counter.fetch_add(1, Ordering::SeqCst); });

    ctx.it("should run after the after", || {
        assert_eq!(0, counter.load(Ordering::SeqCst));
        Ok(())
    });

    ctx.describe("a group of examples", |ctx| {

        ctx.it("should see 1 increment", || {
            assert_eq!(1, counter.load(Ordering::SeqCst));
            Ok(())
        });

        // XXX - note that the after has not been applied another time
        ctx.it("should NOT see another increment", || {
            assert_eq!(1, counter.load(Ordering::SeqCst));
            Ok(())
        });
    });

    ctx.it("should run after the all the afters AND the previous it", || {
        assert_eq!(2, counter.load(Ordering::SeqCst));
        Ok(())
    });
});