Module rspec::header

Headers store the label and name of a Suite/Context/Example.

Structs

ContextHeader
A Header with label and name of a Context.

ExampleHeader
A Header with label and name of an Example.

SuiteHeader
A Header with label and name of a Suite.

Enums

ContextLabel
How the Context will be printed by the Logger.

ExampleLabel
How the Example will be printed by the Logger.

SuiteLabel
How the Suite will be printed by the Logger.