Module rspec::logger

Loggers can be attached to the Runner to log the progression and results of a test suite.

Examples

let logger = Arc::new(rspec::Logger::new(io::stdout()));
let configuration = rspec::ConfigurationBuilder::default().build().unwrap();
let runner = rspec::Runner::new(configuration, vec![logger]);

Structs

Logger
Preferred logger for test suite execution.