Crate rspec

Re-exports

pub use logger::Logger;

pub use runner::Runner;

Modules

block
Blocks are used to build a tree structure of named tests and contextes.

header
Headers store the label and name of a Suite/Context/Example.

logger
Loggers can be attached to the Runner to log the progression and results of a test suite.

report
Reports provide information about an evaluated test unit.

runner
Runners are responsible for executing a test suite's examples.

Structs

Configuration
A Runner's configuration.

ConfigurationBuilder
Builder for Configuration.

Functions

describe
Alias for suite, see for more info.

given
Alias for suite, see for more info.

run
A wrapper for conveniently running a test suite with the default configuration with considerebly less glue-code.

suite
Creates a test suite from a given root context.